# perf: Improve materialisation performance of SortPreservingMergeExec