--- a/datafusion/src/physical_plan/sort_preserving_merge.rs
+++ b/datafusion/src/physical_plan/sort_preserving_merge.rs
@@ -575,7 +575,54 @@ mod tests {
     use tokio_stream::StreamExt;
 
     #[tokio::test]
-    async fn test_merge() {
+    async fn test_merge_interleave() {
+        let a: ArrayRef = Arc::new(Int32Array::from(vec![1, 2, 7, 9, 3]));
+        let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
+            Some("a"),
+            Some("c"),
+            Some("e"),
+            Some("g"),
+            Some("j"),
+        ]));
+        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![8, 7, 6, 5, 8]));
+        let b1 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
+
+        let a: ArrayRef = Arc::new(Int32Array::from(vec![10, 20, 70, 90, 30]));
+        let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
+            Some("b"),
+            Some("d"),
+            Some("f"),
+            Some("h"),
+            Some("j"),
+        ]));
+        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![4, 6, 2, 2, 6]));
+        let b2 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
+
+        _test_merge(
+            b1,
+            b2,
+            &[
+                "+----+---+-------------------------------+",
+                "| a  | b | c                             |",
+                "+----+---+-------------------------------+",
+                "| 1  | a | 1970-01-01 00:00:00.000000008 |",
+                "| 10 | b | 1970-01-01 00:00:00.000000004 |",
+                "| 2  | c | 1970-01-01 00:00:00.000000007 |",
+                "| 20 | d | 1970-01-01 00:00:00.000000006 |",
+                "| 7  | e | 1970-01-01 00:00:00.000000006 |",
+                "| 70 | f | 1970-01-01 00:00:00.000000002 |",
+                "| 9  | g | 1970-01-01 00:00:00.000000005 |",
+                "| 90 | h | 1970-01-01 00:00:00.000000002 |",
+                "| 30 | j | 1970-01-01 00:00:00.000000006 |", // input b2 before b1
+                "| 3  | j | 1970-01-01 00:00:00.000000008 |",
+                "+----+---+-------------------------------+",
+            ],
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    async fn test_merge_some_overlap() {
         let a: ArrayRef = Arc::new(Int32Array::from(vec![1, 2, 7, 9, 3]));
         let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
             Some("a"),
@@ -584,21 +631,92 @@ mod tests {
             Some("d"),
             Some("e"),
         ]));
-        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![8, 7, 6, 5, 4]));
+        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![8, 7, 6, 5, 8]));
         let b1 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
 
-        let a: ArrayRef = Arc::new(Int32Array::from(vec![1, 2, 3, 4, 5]));
+        let a: ArrayRef = Arc::new(Int32Array::from(vec![70, 90, 30, 100, 110]));
         let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
+            Some("c"),
             Some("d"),
             Some("e"),
+            Some("f"),
             Some("g"),
-            Some("h"),
-            Some("i"),
         ]));
         let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![4, 6, 2, 2, 6]));
         let b2 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
-        let schema = b1.schema();
 
+        _test_merge(
+            b1,
+            b2,
+            &[
+                "+-----+---+-------------------------------+",
+                "| a   | b | c                             |",
+                "+-----+---+-------------------------------+",
+                "| 1   | a | 1970-01-01 00:00:00.000000008 |",
+                "| 2   | b | 1970-01-01 00:00:00.000000007 |",
+                "| 70  | c | 1970-01-01 00:00:00.000000004 |",
+                "| 7   | c | 1970-01-01 00:00:00.000000006 |",
+                "| 9   | d | 1970-01-01 00:00:00.000000005 |",
+                "| 90  | d | 1970-01-01 00:00:00.000000006 |",
+                "| 30  | e | 1970-01-01 00:00:00.000000002 |",
+                "| 3   | e | 1970-01-01 00:00:00.000000008 |",
+                "| 100 | f | 1970-01-01 00:00:00.000000002 |",
+                "| 110 | g | 1970-01-01 00:00:00.000000006 |",
+                "+-----+---+-------------------------------+",
+            ],
+        )
+        .await;
+    }
+
+    #[tokio::test]
+    async fn test_merge_no_overlap() {
+        let a: ArrayRef = Arc::new(Int32Array::from(vec![1, 2, 7, 9, 3]));
+        let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
+            Some("a"),
+            Some("b"),
+            Some("c"),
+            Some("d"),
+            Some("e"),
+        ]));
+        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![8, 7, 6, 5, 8]));
+        let b1 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
+
+        let a: ArrayRef = Arc::new(Int32Array::from(vec![10, 20, 70, 90, 30]));
+        let b: ArrayRef = Arc::new(StringArray::from_iter(vec![
+            Some("f"),
+            Some("g"),
+            Some("h"),
+            Some("i"),
+            Some("j"),
+        ]));
+        let c: ArrayRef = Arc::new(TimestampNanosecondArray::from(vec![4, 6, 2, 2, 6]));
+        let b2 = RecordBatch::try_from_iter(vec![("a", a), ("b", b), ("c", c)]).unwrap();
+
+        _test_merge(
+            b1,
+            b2,
+            &[
+                "+----+---+-------------------------------+",
+                "| a  | b | c                             |",
+                "+----+---+-------------------------------+",
+                "| 1  | a | 1970-01-01 00:00:00.000000008 |",
+                "| 2  | b | 1970-01-01 00:00:00.000000007 |",
+                "| 7  | c | 1970-01-01 00:00:00.000000006 |",
+                "| 9  | d | 1970-01-01 00:00:00.000000005 |",
+                "| 3  | e | 1970-01-01 00:00:00.000000008 |",
+                "| 10 | f | 1970-01-01 00:00:00.000000004 |",
+                "| 20 | g | 1970-01-01 00:00:00.000000006 |",
+                "| 70 | h | 1970-01-01 00:00:00.000000002 |",
+                "| 90 | i | 1970-01-01 00:00:00.000000002 |",
+                "| 30 | j | 1970-01-01 00:00:00.000000006 |",
+                "+----+---+-------------------------------+",
+            ],
+        )
+        .await;
+    }
+
+    async fn _test_merge(b1: RecordBatch, b2: RecordBatch, exp: &[&str]) {
+        let schema = b1.schema();
         let sort = vec![
             PhysicalSortExpr {
                 expr: col("b", &schema).unwrap(),
@@ -615,25 +733,7 @@ mod tests {
         let collected = collect(merge).await.unwrap();
         assert_eq!(collected.len(), 1);
 
-        assert_batches_eq!(
-            &[
-                "+---+---+-------------------------------+",
-                "| a | b | c                             |",
-                "+---+---+-------------------------------+",
-                "| 1 | a | 1970-01-01 00:00:00.000000008 |",
-                "| 2 | b | 1970-01-01 00:00:00.000000007 |",
-                "| 7 | c | 1970-01-01 00:00:00.000000006 |",
-                "| 1 | d | 1970-01-01 00:00:00.000000004 |",
-                "| 9 | d | 1970-01-01 00:00:00.000000005 |",
-                "| 3 | e | 1970-01-01 00:00:00.000000004 |",
-                "| 2 | e | 1970-01-01 00:00:00.000000006 |",
-                "| 3 | g | 1970-01-01 00:00:00.000000002 |",
-                "| 4 | h | 1970-01-01 00:00:00.000000002 |",
-                "| 5 | i | 1970-01-01 00:00:00.000000006 |",
-                "+---+---+-------------------------------+",
-            ],
-            collected.as_slice()
-        );
+        assert_batches_eq!(exp, collected.as_slice());
     }
 
     async fn sorted_merge(