@alamb 2021-07-06
I think keeping https://github.com/apache/arrow-datafusion/issues/655 open is fine